New "containers" folder and adrian::FixedList
Created a new container called FixedList which acts like a std::list
but stores it's data statically instead of dynamically (i.e. on the
stack instead of the heap).

This is helpful for systems where dynamic memory is expensive or not
possible, but where there is a need for a container with efficient
insertion/removal in the middle.

An example of this is a scheduling system which uses a sorted list to
store the scheduled tasks. This container was created with a future
system in mind.

Also updated the test script and added tests for FixedList.